Refactor based on Bo's implementation
