+++ .gitignore
@@ -1,0 +1,9 @@
+*.pyc
+
+__pycache__/
+
+wandb/
+
+*.ipynb_checkpoints/
+
+*egg-info/
