+++ rectified_flow/flow_components/train_time_weight.py
@@ -1,0 +1,20 @@
+import torch
+
+
+class TrainTimeWeight:
+    def __init__(
+        self,
+        weight: str = "uniform",
+    ):
+        self.weight = weight
+
+    def __call__(
+        self,
+        t: torch.Tensor,
+    ) -> torch.Tensor:
+        if self.weight == "uniform":
+            wts = torch.ones_like(t)
+        else:
+            raise NotImplementedError(f"Time weight '{self.weight}' is not implemented.")
+        
+        return wts
