+++ rectified_flow/pipelines/train_dit_cifar.sh
